% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_list.R
\name{get_edge_list}
\alias{get_edge_list}
\title{get_edge_list}
\usage{
get_edge_list(df, actor, assoc_actors, idx)
}
\arguments{
\item{who}{I don't know what this is for}
}
\value{
returns a dataframe or tibble
}
\description{
Creates an edge list when the second column list any number of entities for one observation.
}
\details{
\code{get_edge_list()} takes an edge list with an id column where the second column has multiple values. It then separates the second column and creates a full edge list.
}
\examples{
df <- tibble::tibble(
 actor_1 = c("A", "A", "B"),
 assoc_actor_1 = c("B", "B; C", "C; D; E")) |>
   dplyr::mutate(idx = dplyr::row_number()) |>
   dplyr::select(idx, dplyr::everything())

print(get_edge_list(df,
                    actor = actor_1,
                    assoc_actors = assoc_actor_1,
                    idx = idx))


}
