% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_list.R
\name{get_edge_list}
\alias{get_edge_list}
\title{get_edge_list}
\usage{
get_edge_list(df, actor, assoc_actors, idx, sep = ";")
}
\arguments{
\item{df}{dataframe}

\item{actor}{actor or node}

\item{idx}{index, row number or grouping variable}

\item{sep}{separator - default sep = ';'}

\item{assoc_actor_1}{second actor or node with multiple values sep = ';'}
}
\value{
returns a dataframe or tibble
}
\description{
Creates an edge list when the second column lists any number of nodes/entities for one observation.
}
\details{
\code{get_edge_list()} takes an edge list with an id column where the second column has multiple values. It then separates the second column and creates a full edge list. As opposed to \code{separate_rows()}, this function only keeps those pairwise combinations that exist in the original data.
}
\examples{
df <- tibble::tibble(
actor_1 = c("A", "A", "B"),
 assoc_actor_1 = c("B", "B; C", "C; D; E")) |>
   dplyr::mutate(idx = dplyr::row_number()) |>
   dplyr::select(idx, dplyr::everything())

print(get_edge_list(df,
                    actor = 'actor_1',
                    assoc_actors = 'assoc_actor_1',
                    idx = 'idx'))

}
